Changed the way structures are removed.
Added a 'fetchWithRetries' method for the overpass query but can be used for anything.
Also, had problems with downloading "osmtogeojson.js" so I have downloaded it and i then serve it myself. Also also, changed the way that the location watcher works and i understand promises quite a bit more now. Also also also, I have added checks for the new locaiton to ensure it is within the boundaries of the .tiff file.