Improved quality. Added 'quality' buttons
Made the actual map look nicer and added a sun but it takes up 2400 triangles so makes performance worse so i have decided to add two buttons to change the quality of the map. Low quality gets rid of sun and height map. Super buggy tho if you hit the buttons two quickly.