Improved mini-map. Coded out the height map.
Also, added more features to the navigation search auto-complete.
Also also, started trying to get the phones heading and use it to point the mini-map and maybe main-map in the correct direction. I couldn't quite get it to work properly. It wasn't lining up with the actual angle i was giving it.